chart 수정, UI 수정

+++ src/main/java/EHS/MonitoringSystem/vo/ChartDataVO.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class ChartDataVO {
     private String x;
-    private long y;
+    private double y;
 }
